Patch karma-sinon-chai to support modules
At some point, chai.js started to be exported as an ES module, but since
karma-sinon-chai is horribly old, either it didn't know how to handle ES
modules or it wasn't telling Karma how to handle it properly. That lead
to an "unexpected token export" error when Karma was loading chai.js.

To fix that, this patches karma-sinon-chai to tell Karma to load it as a
JS module, and then it imports chai as a JS module as well instead of
expecting chai to exist on the window as it would have in the past.