Add support for React 18 (#4)
* Update repo to support Node 20 and rebuild

This requires us to run node scripts with
NODE_OPTIONS=--openssl-legacy-provider.

* Update to React 17

We never actually updated the library to be built against React 17 when
we updated the rest of MM for it.

* Tests: Fix tests that expected errors

For whatever reason, the updates to React made it so that a mock of
console.error started receiving just the format string for error
messages, so we have to manually format it now.

* Functional: Add UNSAFE_ to deprecated lifecycle methods

I also had to patch the uncontrollable library for this since the newest
version that still includes the HOCs used by this code still uses the
old lifecycle methods.

* Functional: Fix defaultProps warnings for NavbarForm

This isn't actually needed until React 19, but I wanted to do what I can
to get us prepared for that.

* Functional: Remove logic involving string refs

* Test: Remove usage of ReactTestUtils.Simulate in Enzyme tests

* Update React to 18.3.1

* Add React Testing Library

* Update Karma and Webpack

* Functional: Update keyboard handlers to use .key instead of .keyCode

RTL doesn't support .keyCode since it's been deprecated for a long time.

* Tests: Remove explicit usage of ReactDOM.render from tests

ReactDOM.render is deprecated, and using it prevents us from using the
new features in React 18. The same goes for
ReactTestUtils.renderIntoDocument since that uses ReactDOM.render
internally and is similarly deprecated.

* Tests: Remove usage of ReactTestUtils.renderIntoDocument

As mentioned previously, that method is deprecated and uses
ReactDOM.render under the hood.

* Tests: Update OverlayTrigger test to ensure transitions keep working

It seems like removing the deprecated React methods from OverlayTrigger
in a followup commit causes the async parts of OverlayTrigger to run at
a slightly different time. This required me to rewrite the test slightly
similar to the changes I made to CollapseSpec in the previous commit.

* Functional: Replace unstable_renderSubtreeIntoContainer with a portal

This method is deprecated in React 18 (removed in 19), and I think it
prevents us from using the new React 18 features, similar to
ReactDOM.render.

* Tests: Remove test code silencing ReactDOM warnings

* Tests: Replace ReactDOMTestUtils.act with RTL version

* Mark generated files so GitHub doesn't show diffs by default