Patch react-overlays to use KeyboardEvent.key
`KeyboardEvent.keyCode` has been deprecated for years now, and the latest
version of `@testing-library/user-event` doesn't support it. We need to
add support for `KeyboardEvent.key` to be testable in Testing Library.

I made this same change in #4 for React Bootstrap itself, but it turns
out it's also needed in React Overlays as well.