Move patch-package to be a non-dev dependency
This ensures that it's actually downloaded when this package is installed
as a dependency of other code. I chose that instead of making it a
required peer dependency because, while this is a bit janky, I feel like
it makes the library easier to use by MM plugins.

In the long term, we should look at replacing the patched packages or
properly forking them because this is sure to cause problems at some
point because having this package mess with other packages in the
node_modules is a terrible idea. Or we could look at actually publishing
this properly, or we could just remove React Bootstrap from MM finally.
Regardless, this is a hacky fix to keep a hacky fix working.